requirements.txt
streamlit==1.55.0
pandas==2.3.3
numpy==2.4.3
pillow==12.1.1
matplotlib==3.10.8
plotly==6.6.0
folium==0.20.0
streamlit-folium==0.27.1
openpyxl==3.1.5
python-docx==1.2.0
fpdf2==2.8.7
pyshp==3.0.3
pyproj==3.7.1
requests==2.33.0
scikit-learn==1.8.0
ultralytics==8.4.33
fastapi==0.135.3
uvicorn==0.43.0
huggingface_hub==1.9.0
